index.js/root working
