+++ package.json
@@ -9,7 +9,9 @@
   "author": "WDI8",
   "license": "ISC",
   "dependencies": {
+    "body-parser": "^1.17.1",
     "express": "^4.15.2",
+    "express-handlebars": "^3.0.0",
     "mongoose": "^4.4.11"
   }
 }
